refactor(scancode): Move error mapping functions to a separate file
Get a better overview by reducing the amount of code per file.
